requirements.txt
click>=6.6,<7
cookiecutter>=1.4,<2
requests>=2.9.1,<3
requests-mock>=0.7.0,<1
PyYAML>=3.11
python-dotenv>=0.5.1
